+++ nix/cardano-bridge.nix
@@ -1,4 +1,4 @@
-{ target, runCommandCC, cardano-wallet, cardano-node, cardano-shell, cardano-cli, cardano-address, lib, local-cluster ? null }:
+{ target, runCommandCC, cardano-wallet, cardano-node, cardano-shell, cardano-cli, cardano-address, lib, local-cluster ? null, mock-token-metadata-server }:
 
 runCommandCC "daedalus-cardano-bridge" {
   passthru = {
@@ -14,7 +14,30 @@ runCommandCC "daedalus-cardano-bridge" {
   cp -f ${cardano-shell.haskellPackages.cardano-launcher.components.exes.cardano-launcher}/bin/* .
   cp -f ${cardano-node}/bin/cardano-node* .
   cp -f ${cardano-cli}/bin/cardano-cli* .
-  ${lib.optionalString (local-cluster != null) "cp -f ${local-cluster}/bin/local-cluster* ."}
+  ${lib.optionalString (local-cluster != null) ''
+
+    ${if target == "x86_64-windows" then ''
+      # Recursive for selfnode shelley test data:
+      cp -rf ${local-cluster}/bin/* .
+
+    '' else if target == "x86_64-linux" then ''
+      cp -f ${local-cluster}/bin/local-cluster .
+
+    '' else if target == "x86_64-darwin" then ''
+      # For nix-shell:
+      cp -f ${local-cluster}/bin/local-cluster .
+
+      # For selfnode installer:
+      cp -f ${local-cluster}/bin/.local-cluster-wrapped local-cluster--unwrapped
+      mkdir -p test/data
+      test_data_dir=$(cat local-cluster | grep -oP "SHELLEY_TEST_DATA='\K[^']+")
+      cp -rf $test_data_dir test/data/
+
+    '' else abort "Unknown target: ${target}"}
+
+    cp -f ${mock-token-metadata-server}/bin/* . || true
+    cp -f ${./../utils/cardano/selfnode}/token-metadata.json .
+  ''}
   ${lib.optionalString (target == "x86_64-linux") ''
     chmod +w -R .
     for x in cardano-address cardano-node cardano-launcher cardano-cli cardano-wallet; do